requirements.txt
Django==5.2.10
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-filter==25.2
django-redis==5.4.0

psycopg2-binary==2.9.9

django-countries==8.2.0
django-phonenumber-field==8.4.0
phonenumberslite==9.0.22
pillow==12.1.0

python-decouple==3.8

gunicorn==21.2.0

factory_boy==3.3.3
